Validate plugin filters for IGeometry design.
Verifying that the plugin filters correctly data check and preflight for the
IGeometry design.  It appears that most of the dimension checking is done on the
attribute matrices as instead of the data container, so the actual geometry
object is unnecessary.  This may break compatibility with other filters that
expect and image geometry.