First pass a reading a image with ITK
This commit is here to ensure that one can open an image with ITK from
within the dream3d application. It doesn't actually do anything besides
reading the image properties and printing them.
Note that as of now we have to manually register the IO factories from
ITK. In this case, since this commit is just a starting point, only the
PNG factory is registered. Other image format will fail to be opened.