ENH: Following more closely ITK's API
In itkImageToDream#DDataFilter, the data container is embedded
in itkSimpleDataObjectDecorator which allows to really use ITK's API.
A few style improvements have also been done (typedef alignment,
passing arguments by reference, better function naming). Renaming
ValidInputCheck() to VerifyPreconditions() which is a virtual function
in itkProcessObject. This allows to not have to manage the validation
of the inputs manually in itkImageToDream3DDataFilter. Additionally,
GenerateData() was split in two functions: GenerateData() and
GenerateOutputInformation() to replicate ITK's API.