ENH: Specializing ImageContainerTemplated to Dream3DImage
ImageContainerTemplated were aimed to improve the Image class
by letting the developer specify which type of image container
should be used. However, the use case was too limited and
to avoid a complicated implementation, the implementation
was simplified and the template parameter to choose the type
of image container was remove. Dream3DImage only use
ImportDream3DImageContainer which handles memory allocation
with 'malloc' and 'free'.