ENH: itkInPlaceDream3DDataToImageFilter can choose to take data ownership
Addition of a new flag to itkInPlaceDream3DDataToImageFilter to take
ownership of the data or not. The new flag is called
(Set/Get)PixelContainerWillOwnTheBuffer
This is consistent with DREAM3D behavior to store data in Dream3D data containers
while DREAM3D is running. ITK filters are specific to the scope in which
they are running and should most likely never take ownership of the data.
By default, the flag is therefore turned to False.