BUG: Remove comparison to hash for tests
To verify if tests had passed, the hash of the saved file was computed
and compared to the expected value that was hardcoded in the test.
This lead to problems when running the tests on a different plateform
than the one on which the tests were written. To avoid this problem,
tests are now reloading the saved file and comparing its content to
what is expected.

Current limitation include:
-Disabling test on origin and spacing as SCIFIO does not seem to save
or load this information correctly.
-Disabling test on the pixel type. Expected type could be 'int8' and
actual type was 'çhar' which lead to the test failing.