BUG: Verifies data array path in ITKImageWriter
Verifies that the given dataArrayPath actually exists. Using the UI
to select the dataArrayPath to save does not allow to select a
path that does not exist. However, the default path at which the
variable is set may not exists. If one runs the filter without
changing that default path, no check was perform and DREAM3D was crashing.