BUG: Missing Graft() overloaded function
Graft() function had not been reimplemented in itk::Dream3DImage.
When Graft() was called, it is the implementation available in
itk::Image that was called, which lead to filters crashing due to the
fact that it was casting the PixelContainerType to be casted in
the wrong type: ImportImageContainer instead of
ImportDream3DImageContainer.