BUG: Limit number of CPU cores used on CircleCI
When running 'ctest -D Experimental' to compile and test this plugin,
ctest compiles the plugin using all the cores that are available on
the machine, which increases the amount of memory used by the compiler,
and the compilation fails due to requiring to much memory.
To avoid this problem, and without completely re-writing the testing
system and using a CTest script, we just compile DREAM3D firsti, limiting
the number of cores requested for the compilation, and then run
'ctest -D Experimental'.