BUG: On Windows, DREAM3D executable in not in a subdirectory bin
When setting paths for SCIFIO to find jars and jre, only go up one directory
if DREAM3D is running on a "not Windows" system. On these systems, the
executable is installed in the 'bin' subfolder whereas on Windows the
DREAM3D executable is installed directly in the main DREAM3D folder.