ENH: Add filterCastToFloat method in ITKImageBase
This is used by filters that require floating point pixel types to operate
well. The filter should be templated over FloatImageType.

This path casts the input to itk::Dream3DImage< itk::NumericTraits<
InputPixelType >::RealType, Dimension > before passing it to the filter, then
casts to the OutputImageType on output. Note that the CastImageFilter is a
no-op when its input and output have the same type.

Issue #59