BUG: Use filter_type value only when not empty
filter_type is used to define the type of the filter if it is not
the expected type (FilterName<InputImageType,OutputImageType>).
When parsing the JSON file and the dictionaries to verify the keys
available, filter_type is always initialized to an empty string and
therefore the key is always in the dictionary. We need to check
its value to know if we should use it or not instead of only checking
if the key is in the dictionary.