BUG: The number of components of the output image is based on the output pixel type
Previously, in the preflight step, the number of components of the output image was
computed based on the input pixel type. This addresses [1].

WARNING: If a filter uses input and output pixel types that have different number
of components, a new array is required. The boolean checkbox to select if a new
array should be used should be hidden. By default, this boolean variable is
set to 'true' in ITKImageBase constructor.

[1] https://github.com/BlueQuartzSoftware/ITKImageProcessing/issues/114