ENH: Support JSON with 'measurements' and 'measurements_results'
When a SimpleITK JSON file contains the keyword 'measurements' in the description
or the filter or the keyword 'measurements_results' in the description of the
test, the filter may have an additional output that is a value. This value
needs to be printed in DREAM3D GUI and accessible to the test (it is set as
a property of the filter).