Create ITKImage library that can be used in other plugins
When creating filters that are based on ITK in other plugins, it is convenient
to be able to use what has been implemented in this plugin for this purpose,
and not have to duplicate this code. This patch creates a library that
can be integrated in other plugins for this purpose.