New macro to support both 2 inputs image types with different component types.
`Dream3DArraySwitchOutputComponentMacro()` is a new macro integrated in
`Dream3DTemplateAliasMacro.h` that allows to automatically select the second
template type for the templated functions `filter()` and `dataCheck()`, based
on the component type of a data array. This allows the user to develop filters
that take 2 inputs that can have different component types.