Resolves nathanblair/DREAM3D#77
Also resolves nathanblair/DREAM3D#76

Begun work on nathanblair/DREAM3D#75

Current progress indicates something is off in retrieving the maximum
value from the output FFTConvolutionFilter. The residual being set is
on the order of 10^137...which seems extremely high for an initial guess
that is precisely the same as the correct transformation matrix.